moved coroutinescope example to chapter 2
