+++ Tutorial1-1CoroutinesBasics/src/main/res/layout/activity2_scope_2.xml
